CMS-13367 [Backport 14.3] Always disable content toggle if project is not editable
Use function references to move the logic of when to initially render
a disabled overlay button from the OverlayController into the
ChannelController.

The isComponentOverlayInitiallyDisabled can now be re-used for the
state of the components tab in the left side-panel.

(cherry picked from commit 8036d88ac112e6a416138cd764ffec5c9c811281)