CMS-13284 [Backport 13.4] Check if violations for this field exist before re-validating
When adding, moving or removing a content-block item, we trigger
validation in order to keep the state of the document correct. For
example, when we have a block with a validation error and we remove
that block, it's violation message should be removed as well.

Prior, we would always trigger validation of the entire document upon
adding, moving or removing a content-block. This is not an ideal UX.

This fix tries to make it more subtle. Only trigger validation (of the
entire document) if there are existing violations in/below the
content-block. This means that the UI will stay in sync when adding,
moving or removing content-blocks with existing violations, but these
will only be visible if the user already pressed "save" once.