CMS-13850 Only add components that belong to page
When rendering a page Wicket expects all components to belong to the same page
instance. In Wicket versions prior to 7 only a warning was logged, but from 7 on
the framework throws an Exception when this happens. As a consequence an e2e test
started failing.

It seems this problem exists already for a long time because the class
OnlyRenderComponentsOnPage was introduced to prevent this from happening.
The fix worked as long as there was only one version of the PluginPage.
However, with the introduction of the navapp there are now 2 versions of
that page so the "fix" no longer works. This commit re-establishes the intent
of the OnlyRenderComponentsOnPage class.

The preferred solution however would be to fix the root cause, but after long
debugging sessions we did not manage to find out how a component can end up
on the wrong page. Maybe it has to do with the lazy loading of plugins?

(cherry picked from commit 2d79afa752e81c260a0ffd617d612f97a7b7c30e)