CMS-13936 [forward port 14.4] Avoid concurrent modification exc and correctly handle referenced components
1) the target list was checked whether it contained 'dynamicParameter.getName()'
however the List contains DynamicParameter so never a match. The problem now
got really big if the component config and the referenced component config
both point to the same HstComponentClass: In that case, both source and target
are the exact same instance, and then in the loop when

   target.add(dynamicParameter);

is done, a concurrent modification exc happens since the source is the same
instance and gets modified as well. This issue fixes this problem and the
first added integration test method covers this

2) in CMS_13877, component parameter name/value inheritance was by accident removed,
however, dynamic component parameters and name/value of stored parameters
should be treated separately. Restored this, the last integration test method
covers this

3) Conceptual merging of DynamicComponentParameters was a bit flawed in case the
referenced component has a different ParameterInfoClass. Covered for this including
an integration test

(cherry picked from commit b9b4cb15419f5b5cebe2ffcc9c25b5efba09f345)