CMS-13997 Make sure to fetch the value via 'residual parameters'
componentParametersInfo.getResidualParameterValues() holds values which
take targeting *and* potential value-override-by-query-param in CM into
account. This is opposed to getComponentParameter(param.getName()) which
does not take this into account.

Also note that non-residual document getter from a parameter info
interface should never be needed to automatically set as model. See the
jira ticket for an elaborate explanation in the description