CMS-13138 [Backport 14.4] NavLocation has ADD or REPLACE mode
The "mode" field should not be used in the equals or hashcode
functions.

If a NavLocation is in ADD mode, the location will be appended to the
history.
If it is in REPLACE mode, it will replace the current location.

(cherry picked from commit af92e5de5f552ef8027d69b161879bd9dd0c7895)