CMS-13138 [Backport 14.4] Select parent when closing last tab
When closing the last document tab, we want the parent folder to be
selected and shown as the current nav-location.

This was handled by the EditorManagerPlugin, by actively setting the
parent node as the current node, but this causes an issue when we want
to close the current document-editor from the channel-manager:

- start editing a document in content-perspective
- now start editing the same document in the visual editor

This sequence causes the system to try and close the
content-perspective editor, which leads to a redraw of the
content-perspective. But this redraw should only become visible once we
navigate to the content-perspective, not while we are in the
channel-manager. So instead of setting the parent folder in the
EditorManagerPlugin (which causes the content perspective to receive
focus) we now handle it in the BrowseState itself.

(cherry picked from commit 50314c3bd3da46d00d15411d310202bc536fed43)