CMS-14040 [Backport 14.4] Simplify TraceMonitor usage
The TraceMonitor is a developer tool that is heavy on resources, but
it can/should also be very dumb (as it should only be used on a
developer machine after noticing warnings about "Undetached
JcrItemModel".

In that sense, it feels more appropriate to simply store the
construction stacktrace in the JcrItemModel when the TraceMonitor is
enabled, so we can log it when an "Undetached JcrItemModel" is found.

The functionality of the TraceMonitor is reduced to:
 - reading a System property to detect whether tracing is enabled
 - proving a utility method that returns a String representation of
the current stack.

The unit test did not add any value anymore.

(cherry picked from commit 834ab13260c435319c68f488fe10e8f5c4494875)