CMS-14022 [Backport 14.4] Support *parsed* spaUrl via Channel object
There was a problem with the previous approach: Only request time in
HstDelegateeFilterBean the previewUrl would be parsed for placeholders.

As a result, the property in the channel manager frontend code would
still use:

org.hippoecm.hst.configuration.channel.PreviewURLChannelInfo_url

which was the unparsed property, resulting in breaking the FE to not
show the channel settings, see CMS_14039 where this must be improved.

Therefore, the new approach is to add to the Channel object the getter

which contains the parsed value for PreviewURLChannelInfo_url. There is
a name mismatch with 'spaUrl' but it is really only used for SPA's so we
keep property PreviewURLChannelInfo_url for 14.x BC but will rename that
in SaaS later.

Also changed the PropertyParser to be forgiving wrt unresolvable place
holders : this way at least it can be seen that the intention was a
redirect to SPA...otherwise getSpaUrl would be null and a normal HST
rendering would be done which is not desired

(cherry picked from commit 4955140a7769c2ba2503be76ca70d7f72b7d4e20)