CMS-13680 Allow custom translation files
To ensure a locale like 'zh_CN' would properly load the 'zh.json'
translation file, we use the aliases argument of method
$translationProvider.registerAvailableLanguageKeys(keys, aliases).

Previously, we simply defined all builtin languages and added aliases
for them, but this means custom languages will not available. Because we
don't support switching languages while the application is running, it
does not really make sense to define all builtin languages upfront.

Instead, we should just use the language that is passed in the
ConfigService, and if that language it not available, use the FALLBACK
language.

I also added error logs to help developers follow what happened.