CMS-14066 Bootstrap elastic7 client
We use the elasticsearch-maven-plugin to automatically start ElasticSearch
when Trends and Experiments are added as an Essentials addon. This plugin
is configured to start ElasticSearch 7. Since we can't change the default
delivered with targeting bootstrap (would auto-update existing projects to
elastic7) the bootstrapped targeting elastic client is elastic6. So we end
up with incompatible versions. Therefore add yaml that will automatically
bootstrap elastic7. This will also ensure that new projects will start with
ElasticSearch7 by default.