CMS-14066 replace visit-store.yaml with visits.yaml
I used autoexport to generate the desired config change so that
it is close to what would happen when a developer would do the
change as documented. But then I forgot to add the file to git
while I did update the file name in the instruction set xml file.
So locally all was fine but of course it would fail when using
the deployed artifacts built by Jenkins.