CMS-14095 Make sure that the callback PMA URL param is added automatically
[Backport of CMS_14086 with adjustments for PaaS version]

With these changes, when a redirect is done to an spaURL, the HST makes
sure to automatically inject the query param 'endpoint=....' where the
param name 'endpoint' is configurable through hst config props however
most likely never needed to be changed.

The value of the endpoint URL points to the *current* (LIVE) channel +
the hst:pagemodelapi postfix. So assume the Channel Manager requests the
following iframe URL

http://localhost:8080/site/_cmsinternal/api/brxsaas

and assume the mount at /api/brxsaas has:

- spaURL = http://localhost:3000
- hst:pagemodelapi = page

then the iframe will be redirected to the URL

http://localhost:3000?token=xyz&server-id=x&endpoint=
http://localhost:8080/site/api/brxsaas/page

Note that thus endpoint is equal to

1) the request channel URL in the brx environment without the _cmsinternal
2) + appended is the value of hst:pagemodelapi since that is where the
PMA mount is located

(cherry picked from commit de42829fe29d6da7097195189bd74014546e36c3)