CMS-13768 [Backport 14.4] Add WorkflowRequestTypes CANCEL, ACCEPT, REJECT and REJECTED
The first request in the "request" hint is used to determine which
HstActions should be exposed to the frontend.

If there is a REJECTED request, we render the default (un)publish
actions. If there are ACCEPT/REJECT/CANCEL requests, we hide the
default publication actions.

(cherry picked from commit dd2fd9b674374390f4f548cb9aa910eb7896a05d)