CMS-13680 Ensure fallback language is loaded
I expected the $translationProvider to automatically loaded the data
of the fallback lange, but that is not the case. By explicitly loading
the fallback language first, we ensure that it is used as expected; if
a key does not exist in the current language, use the one in the
fallback language.