CMS-13768 [Backport 14.4] Move DocumentWorkflowInvokerPlugin to cms-workflow-frontend
The DocumentWorkflowInvokerPlugin requires access to the Request object
to differentiate between ActionDescriptions of type "cancel" and type
"rejected". Therefore, we move the files to the cms-workflow-frontend
module and add a dependency on the workflowmenu module to access the
AbstractWorkflowManagerPlugin, ActionDescription and MenuHierarchy.

(cherry picked from commit a193f1c86c8cbbbc4bb3f9f670492092b15dade7)