CMS-13768 [Backport 14.4] Move DocumentWorkflowInvokerPlugin to cms-workflow-frontend
The DocumentWorkflowInvokerPlugin requires access to the Request object
to differentiate between ActionDescriptions of type "cancel" and type
"rejected". Therefore, we move the files to the cms-workflow-frontend
module and add a dependency on the workflowmenu module to access the
AbstractWorkflowManagerPlugin, ActionDescription and MenuHierarchy.

(cherry picked from commit 631bc4ba8046bf5383cbab58205fb3574e7bf7eb)