CMS-13746 [Backport 14.4] Add publishableBranch action
Uses can perform publish branch action if
- user is granted hippo:editor privileges on the handle
- AND an unpublished variant exists
- AND the unpublished variant is modified
- AND the branchId provided in the call is contained in documentHandle.branches
- AND the document is NOT transferable

So, if we always provide the branchId (either 'master' or branchId from project)
when getting the hints then publishBranch can be used to activate the "publish"
button for Experience Pages. With this change in the back-end it should be possible
to implement that in the front-end.

(cherry picked from commit 805b6da2036c34cd052c0358e7140cb24fe15834)