CMS-13746 [Backport 14.4] Replace ensureEditorIsPristine() with reloadEditor()
If the editor is open and dirty, we simply call a reload to trigger the
save/discard dialog (no need for ensureEditorIsPristine()) before
executing the workflow action. This has the added benefit that the
editor is already in the  expected state when a dialog from the workflow
itself is cancelled (like with take offline).

After a successful workflow action we always reload the editor to
reflect the  new state.

(cherry picked from commit 511d3ef411f93b9f0e1207334b7203f0bdd8624d)