CMS-13060 Add text search to taxonomy picker
The text search uses https://github.com/robert-bor/aho-corasick to
match categories based on their (localized) name, description and
synonyms.

When a category is matched, it is displayed and its ancestor will be
rendered greyed-out. When a category does not match and has no
matching child categories, it is hidden.

The tree will be expanded to highlight all the matching categories.