CMS-14147 [Backport 14.4] Add "browse" event handler to facilitate copy/move workflow
The existing workflow actions for copy&move open up a Wicket browse
dialog to pick the destination and name. On success, the dialog will
invoke a browse event with the copied or moved document model.

The DocumentWorkflowInvokerPlugin now registers itself as a
IBrowseService which is configured in channel-manager-cluster.yaml.

When a browse event is triggered by the workflow dialogs, the ID of
the node is stored in the browser as the workflow payload. Once the
dialog is fully closed, the workflow will be resolved with the payload
that was set by the browse event, in this case the document ID, which
will be used to navigate to the new xpage in the CM.

(cherry picked from commit baed2620aaf289fa536e32bc3f136175a7f0fd63)