CMS-14147 [Backport 14.4] Add copy, move and delete menu items for xpages
After the copy&move workflow actions are resolved, the URL of the
new page is retrieved from the server and used to navigate
the iframe to it.
In case of delete, a successful workflow action will load the first page
of the sitemap.

(cherry picked from commit 3785cd65cc08ceeb1921b1eb2a9ab9e8c56f07fb)