CMS-14147 [Backport 14.4] Add copy, move and delete menu items for xpages
After the copy&move workflow actions are resolved, the URL of the
new page is retrieved from the server and used to navigate
the iframe to it.
In case of delete, a successful workflow action will load the first page
of the sitemap.

(cherry picked from commit 7574f073b42634564c658f6cac9e0262b0436fa1)