CMS-13694 [Backport 14.4] Support Hst Component structure next to containers in XPage Docs
Functional requirements

To an XPage Doc it should be possible to add also static components &
containers, besides to what was the case, that only containers with a
UUID as name which linked back to a container in the XPage Layout were
supported. So the goal is that for example the following should also be
supported:

hst:xpage [hst:page-ref = two-column]
 + f779f72a-a626-4f3f-869f-2f2c9537bc6c
 +  73711d4b-f1cb-4f4b-b874-69c6ce2b9f0a
 + a87d07c9-bba9-439b-990f-4368bd500094
 + c252dc09-1086-4b75-8130-c41be6e980c2
 + main
    + main/extra
       + main/extra/search-bar

In the above example, the 'main/extra' should be merged into the
XPage configuration for the xpage doc. Assuming two-column page-ref
already has 'main', the 'main/extra' needs to be merged in, where
the 'main' from the two-column page layout has PRECEDENCE wrt
the same defined properties. Properties only on the XPage Doc 'main'
will have to merged in. Same goes for descendant nodes.

For the CM UI, it is crucial to 'know' whether a component is from
'shared' config or not. In the example above, 'main' should be marked with

shared = true

since it comes from the XPage Layout (or even from a base page which
the XPage Layout page references). The /main/extra should be marked
with

shared = false

assuming the XPage Layout does not have '/main/extra'

Key implementation details

The implementation has been implemented completely test-driven, where
first in hst-unittestcontent.yaml an XPage Doc is bootstrapped containing
extra HST component configurations and in ExperiencePageIT the expectations
are set up, where by far the most tricky part was to get the following right:

1) The inheritance precedence
2) The shared true|false flag
3) The isExperiencePageComponent flag
4) The isXpageLayoutComponent falg

These flags are super important to get the CM UI right: If for example the flag
shared is not set right, then for example only a cms webmaster can modify the
container, which of course would be incorrect if the container is defined in
the XPage Doc itself

All the meat of the coding details are done in ExperiencePageServiceImpl and in
HstComponentConfigurationService : the code is not so hard, the functional
details are the hard part, hence everything is cast in concrete in
integration tests asserting all the exact implemented behavior.

The added code in ExperiencePageServiceImpl could be extracted into a separate
method : admittedly the ExperiencePageServiceImpl#loadExperiencePage method
is becoming a bit large. Extracting a method however I am not sure will really
improve readability, however with all the integration tests in place easy to do
now

(cherry picked from commit b01831f60799ee1b302447f5b854ec51bcf147a3)