CMS-13581 [Backport 14.4] Support attributes to indicate unresolved XPage layout container
when the channel manager adds a new container item to a container that
is not yet present in the XPage document, we need 3 pieces of information

- The identifier of the hst:xpage node below the DOCUMENT VARIANT
- The identifier of the XPage Layout container id
- The identifier of the Catalog Item to be added

with this information, the Channel Mgr can construct a somewhat different
RESTY call to inject a catalog item in an unresolved XPage layout
container

For this reason, when there is a Container XPage Layout which is *not*
accounted for in the XPage Document, then in CmsComponentComponentWindowAttributeContributor
we make sure that the following attributes are added to the html comment:

HST-Experience-Page-LayoutComponent = true
HST-Experience-Page-Layout-Hippo_Identifier = {hippo_identifier}

The 'uuid' of the container might be confusing, since it is not really the
'uuid' of the container any more, since the XPage Doc does not have the
container. Therefore, the UUID to operate on is the UUID of the hst:xpage
below the xpage document variant. It might be cleaner to call it pageUUID
and to not have a uuid attr at all but this might become too impactfull for
the UI code.

As a result, the UI now needs to behave as follows: Normally, when adding
a component item to a container, the UI does a POST to

/_rp/{uuid}./{catalog_uuid}

where the 'uuid' is the 'uuid' of the container

In the case when the XPage doc does not yet have the container yet, aka when
the attribute

HST-Experience-Page-LayoutComponent = true

is present, then the call must become

/_rp/{uuid}./{xpage_container_hippo_identifier}/{catalog_uuid}

All the behavior is backed by integration tests including full request
tests mimicking the channel mgr invocation

(cherry picked from commit e13d4451c12ff1689dc468d3236fc202e9947e28)