ENT-2172 Add "switch-to-project" state for xpages
When finishing a new xpage, we should transition to the
"switch-to-project" state when viewing a project variant. This behavior
already exists for documents.

After we successfully switch to a project, we should reload the page.

(cherry picked from commit 02b8a19e53d6699ca460d49f06449ec8c0b05d80)