ENT-2139 [Backport 14.4] Support removing actions and states
Make it possible for a ActionState provider to remove an action.
We can do this by making enabled a Boolean field. If the field is
null and the action was already present in the action set then it
is removed, otherwise the existing value is replaced.

Although not strictly needed I decided to do the same for the State.
Note that the implementation for equals/hasCode of Action and State
are now kind of awkward, but it's needed to be sure we can remove an
entry from the HashSet.

(cherry picked from commit dc1ad8c78022c0512f8383aa5875149ffde99914)