ENT-2139 [Backport 14.4] Use maps instead of sets
The implementations for Action and State implemented equals and
hasCode because they were collected and merged in a set. These
implementations were a little awkward because they did not include
all fields. Then I realized that Action and State are only needed
for the JSON representation, so internally I could use a different
structure. I also realized is that the fields category and name are
actually the unique keys for Action and State. So internally it is
easier to use maps. The keys of these maps are implementations of
interface NamedCategory. By letting the existing enums implement
NamedCategory we get equals and hashCode for free.

(cherry picked from commit f4fa2cf806b03f715f8dbe7a11b0b63bf06c0ad6)