ENT-2170 Fix the actions and state and add integration test assertions
There were two wrong assumptions before that needed to be corrected:

1. Unpublished in XPageContextFactory could be for wrong branch

The code was:

final Node unpublished =
          userSession.getNodeByIdentifier(contextService.getExperiencePageUnpublishedVariantUUID());

This always returns the JCR workspace unpublished, while the intention
was that it would return the actual branch (if present). That is not how
it works. As a result, when a branch was created, say foo, then the
workspace version of unpublished would be for 'foo' : So even if the CM
showed 'core', then the XPage unpublished of 'foo' was used. Of course
not the intention

2. In DocumentStateUtils the code

getPublicationStateFromHandle(final Node handle)

Does not take branches into account: This meant, always the doc state
of the JCR workspace unpublished was used. For example the master
version would show state 'changed' if, say, branch 'foo' got modified
and was still the JCR workspace unpublished.

Ideally, this info would come from workflow hints instead of pulling it
ourselves of the document variant. Since modifying the hints is more
work and in 14.4 we are not allowed to modify the scxml, instead use
BranchHandleImpl since this does the job for getting the right JCR
unpublished variant (from workspace or version history)

There are some TODOs to validate, but this is mostly to do functional
validation whether some states are desirable or not

(cherry picked from commit 608788b82c541cc2d4a6f94fe4b85dc79557f408)