ENT-2170 in XM only allow xpage-copy if selected branch exists
The user making the copy most likely expects that a copy in the context
of a selected branch will also belong to the same branch. We therefore only
allow copy if the branchId of unpublished equals the selected branchId. So
the copy action for an xpage is enabled if there already exists a branch for
it, either under the handle or in version history.

Turns out that in SCXML copying a document is always allowed if there is an
unpublished variant even if it is a branch (i.e.g belongs to a project). The
implementation of copy for a branch is such that it copies it without creating
a master version first. Our assumption however was that all documents always
would have a master version. If not things might break (although I cannot name
a concrete example). After lengthy discussion we decided that we're going to keep
it this way and not change SCXML because:
- This has already been the case since the introduction of projects. There are
  no jira issues reported by customers that have this as a cause. Could be that
  customers never copy documents while looking at a branched version could also
  be that they don't report it.
- We have to backport the feature to 14.4 so we have to avoid SCXML changes as
  much as possible.
- For SaaS we'll most likely have to support copying a branch without having a
  master version of it anyhow.
- Although copy is allowed according to SCXML we can disable it in the HST backend
  code, so then there is at least no risk that our production code will do the
  wrong thing.

Since you can also copy a branched xpage via the Content Editor we should also change
the backend code there, but that is quite difficult. We decided that instead of fixing
it we'll file it as a Known Issue in Jira.

(cherry picked from commit 8b919932eeccfcab76cfa2a59c72ecbf48949e0b)