ENT-2170 Disable powermock dependent tests
When XPageResourceTest runs after another resource test
then it fails with a ModelRegistrationExceptio with a
RepositoryException root cause:

the repository home /some/path/to/target appears to be in use since the file
named .lock is already locked by the current process.

Why this happens is unclear to me. For now we ignore the tests and add a TODO
to fix it later.

(cherry picked from commit 6c2344464d1347f40e334fc4435d9fe4ea0c0e6e)