CMS-14112 Support configuring custom root xpage folder
Customers can change the root XPage folder by adding the mixin
hippostd:rootxpagefolder. For each channel you can set this mixin
on one of the child nodes of it's channel content root. When the
channel state is requested in the context of an xpage then the
ChannelContextFactory will first try to find the custom xpage root
folder, otherwise it will try to find it on the direct children
of the content root. When found (either custom one or as direct child)
it will be used to get the available template query for the channel.