CMS-14112 Rename mixin and let it inherit xpagefolder
The mixin is meant as config for front-end purposes, e.g.
to configure where to start looking for xpages. In addition
we should also support more than one such xpage subfolder.
So the name rootxpagefolder is rather misleading. Since it's
purpose is now for picking xpages in the channel manager the
suffix cm looks ok.

Letting hippostd:cmxpagefolder extend from hippostd:xpagefolder
will make it clearer that these two mixins are interrelated and
that the cmxpagefolder implies the xpagefolder mixin and all of
it's declared properties/nodes. Now the query should no longer
include hippostd:xpagefolder because of the inheritance.