CMS-14286 [Backport 14.5] Fix tests depending on order of execution
for some vague reason, the order of the tests resulted in
failures on jenkins but not locally. It turned out that the
prototyped request from spring component had a different
hostname if ContainerURLProviderIT test ran before it and
setting a hostname on the mock request : It is really vague
why this would impact the mocks created by spring component which
have scope 'prototype'. any way, there is an easy workaround to
not use Spring components to create the mocks

(cherry picked from commit 7a2fb07022e376e69597d56708e3ba46430a6b27)