CMS-14330 Close streams before writing to temp file
When writing to a temp file, a move operation is executed from the internal file that the FileUpload object is using to the new temp file that will be returned. In linux there is no problem, however, in Windows, this move operation fails if we don't ensure the streams are closed beforehand.