CMS-13365 [Backport 14.5] Use Wicket workflow for publish from VE
Because this actually uses the 'publishBranch' workflow action the
TodoPlugin translations need to be aware of it.

To enable the 'cancelRequest' button in the VE panel, the HintsInspector
has to be aware of 'publishBranch as well.

(cherry picked from commit 755fb4a24a47704d51613c215d47e7ea0c7c58ce)