CMS-13365 [Backport 14.5] Use Wicket workflow for publish from VE
Because this actually uses the 'publishBranch' workflow action the
TodoPlugin translations need to be aware of it.

To enable the 'cancelRequest' button in the VE panel, the HintsInspector
has to be aware of 'publishBranch as well.

(cherry picked from commit 87584b4ed799d73109b7909a6d0a8fc867c1a4c1)