CMS-13845 Always supply filename as parameter
ResourceHelper#setDefaultResourceProperties eventually is called by both
the "Add file" option in the foldermenu of an asset folder ( initial
upload) and the "Edit" option in the menu of content editor of a
resource (replace uploaded file by a new file upload).
The DefaultGalleryProcessor did not include the filename as argument.
The property was not set if the filename was null or empty.
The setDefaultResourceProperties without filename argument called the
setDefaultResourceProperties method with filename and the filename
argument was null.
The DefaultGalleryProcessor is eventually used by the UploadDialog.
The UploadDialog is used by the "Add file" option in the foldermenu.
So when a resource was created for the first time, the hippo:filename
property was not set.

Mind that this did not impact the functionality in the cms, because
ImageDisplayPlugin#lookupFilename uses the node name if the property has
not been set. However the delivery tier always uses the hippo:filename
property ( see HippoResource#getFilename and
HippoResourceVisitor#visitNode ), so after initial upload the original
filename of the asset was not accessible.

The filename is now supplied as an argument, so it will be set.

Remove the method signatures without the filename so using the variant
with filename is enforced. Since filename is always present ( and should
be non-null and non-empty ) the null and empty check can be removed. To
follow the style of the code of the class and other cms code I omitted a
null check and a log warning or throwing an exception.

(cherry picked from commit 6a0f194f5ea8113a0d467a60d1e06dad13928424)