CMS-14373 Make the methods in the ParentApi non-optional
The practise is that these methods are defined, right now, and that
child apps can depend on this fact.

The idea that these methods should be optional originated from the
requirement that a parent app could potentially deal with a child app
with a higher, backwards incompatible, version of the nav com lib than
the parent app itself. E.g. we could have added a method to the
ParentApi that older versions of the nav com lib do not know about. It
would be the child apps reponsibility to check for the existence of the
ParentApi method. However this scenario is currently non-existent. When
we would reach that point we should simply add multiple versions of the
parent api interface so child apps (if they use typescript) can use that
as help and base their calls on the version communicated by the parent
app.

(cherry picked from commit 3128702cffd0a9565deeecf9419c5b8b4b9b7d98)