CMS-14209 Traverse target and source folder
In the while loop, the source tree was traversed up, but the target tree
wasn't.
Foldertranslations were added to the folders attribute that all
contained the same target folder, the lowest folder in the tree.
As a result, the translation of the lowest folder in the tree was
repeated in the ui.
Let TranslationWorkflowPlugin#findTranslatedTargetAncestor return the
translated folder.
Make the method recursive to prevent side effects ( modification of
source and target translation folders ).
That way the target and source folder "in sync".
I tried to keep the changes as minimal as possible to not introduce
additional regression.

(cherry picked from commit b3356967da8e75c98eb8b89e5996c46f8474cb62)