CMS-14290 [Backport 13.4] Always reindex a handle or document JCR Node
'reindex' in this context means: first try to remove it from the index
and then index it again

'index' in this context means: index it without first attempting to
delete it

In general, for newly created JCR Nodes, Jackrabbit never does a
'reindex' but always an 'index' : this is because the JCR Node when
added newly has never been indexed before

*However*, for a Hippo Document and Handle, we index a bit more enhanced,
and as a result, in a clustered setup, this can result in a new JCR Node
to be added twice to the index: Once as part of the Journal Event for a
new handle node and once as part of the Journal Event for the new node
itself.

It is due to the added Handle node through which the children are
indexed already that this exceptional case can occur: see the jira
ticket for the exact flow

This commit contains the super simple code fix from '&&' to '||'

The ClusteredIndexingDocumentsTest is a clustered integration test which
*fails* consistently without the fix. The test mimics how workflow
interacts with JCR Nodes. Unfortunately, due to class loading challenges
running real workflow instead of direct JCR manipulation is not possible
in the current cluster setup

(cherry picked from commit 3877cfe406c0f4490bc86d96734d50829a133c8e)
(cherry picked from commit 014b1cc003c9d47a5fda8ed8e38f2550baba0c42)