CMS-14035 Archive handle if only draft exists
Archive the handle if only a draft variant exists.

Cause:
If only a draft variant existed
then there is no unpublished variant
so the archive call was never made.

Fix:
Only remove the draft variant if there is also an unpublished variant.
Use the unpublished variant and otherwise the draft variant to obtain
the workflow.

There is no need for a null check as there ALWAYS is either a draft or a
an unpublished variant. Also a document is not shown in the ui if it has
no variants. If the document only has a published variant, if has be
depublished before deletion, so a unpublished variant is created.
Upon creation the user can either save the draft ( a draft variant is
present ) or saves the document in the "traditional" way and an
unpublished variant is created.