ENT-2375 [backport-13.4] Bump to httpclient 4.5.13
Prior to version 4.5.7 of httpclient empty path segments were removed. But as of 4.5.7 that
is no longer the case. Due to the way the jest client constructs the URL this will cause
issues when the server URL used by the client ends with a slash. The formal solution would be
to create a PR and fix it in jest-client. But since that project is dead it is very unlikely
that a new version will be released. So, instead I fixed it by removing all the trailing slashes
from the server URL locations. This is a better option than asking all customers to change their
server URLs in their configuration.

(cherry picked from commit bdca6c1ad8a57c1909a653a2a0f890134e44484b)