CMS-14385 Encode the lastPathInfoElement
Encode the lastPathInfoElement.

Analysis:
HTTP Header fields parameters cannot carry characters outside the
ISO-8859-1 character encoding. [1]
If there are angular gives the following error: [2], See [3].
A user can enter any characters in the lastPathElement input field.
On any error a toast is shown.
So non-ISO-8859-1 ends up in the header
and a toast is shown and the page is not copied.

Solution:
Use the percent encoding as described in C.2 of [1]. See [4]. I have not
found it in the document, but tomcat supports this encoding.

[1] https://tools.ietf.org/html/rfc6266#section-5
[2] TypeError: Failed to execute 'setRequestHeader' on 'XMLHttpRequest': String contains non ISO-8859-1 code point
[3] https://plnkr.co/edit/SYbX0slJUHDwHccp?open=lib%2Fscript.js&preview
[4] https://tools.ietf.org/html/rfc3986